Fix conflict with `jest.mock('fs');`
