+++ packages/tests/jest.config.js
@@ -12,5 +12,5 @@ module.exports = {
     testEnvironment: 'node',
     testMatch: ['**/*.test.*'],
     roots: ['./src/'],
-    setupFilesAfterEnv: ['<rootDir>/src/setupTests.ts'],
+    setupFilesAfterEnv: ['<rootDir>/src/jest/setup.ts'],
 };
