+++ packages/tests/src/jest/toRepeatStringRange.ts
@@ -2,31 +2,31 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { truncateString } from '@dd/core/helpers';
 import type { MatcherFunction } from 'expect';
 
-export const toRepeatStringTimes: MatcherFunction<[st: string, occurences: number]> =
+export const toRepeatStringRange: MatcherFunction<[st: string, range: [number, number]]> =
     // `st` and `occurences` get types from the line above
-    function toRepeatStringTimes(actual, st, occurences) {
+    function toRepeatStringRange(actual, st, range) {
         if (typeof actual !== 'string' || typeof st !== 'string') {
             throw new TypeError('Only works with strings.');
         }
-        if (typeof occurences !== 'number') {
-            throw new TypeError('Need a number here.');
+        if (!Array.isArray(range) || range.length !== 2) {
+            throw new TypeError('Need an array of two numbers for "range".');
         }
 
+        const { truncateString } = jest.requireActual('@dd/core/helpers');
         const result = actual.split(st).length - 1;
-        const pass = result === occurences;
+        const pass = result <= range[1] && result >= range[0];
 
         const time = (num: number) => (num > 1 ? 'times' : 'time');
         const failure = !pass
             ? `\nBut got it ${this.utils.printReceived(result)} ${time(result)}.`
-            : '';
+            : '.';
         const expected = this.utils.printReceived(truncateString(actual).replace(/\n/g, ' '));
 
         const message = `Expected: ${expected}
 To repeat ${this.utils.printExpected(st)}
-Exactly ${this.utils.printExpected(occurences)} ${time(occurences)}${failure}.`;
+Between ${this.utils.printExpected(`${range[0]} and ${range[1]}`)} times${failure}`;
 
         return {
             message: () => message,
